PODS-1525 : added view for PSA Id and done package restruchuring
