+++ app/views/invitation/psaName.scala.html
@@ -1,0 +1,47 @@
+@*
+ * Copyright 2018 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *@
+
+@import config.FrontendAppConfig
+@import uk.gov.hmrc.play.views.html._
+@import controllers.invitation.routes._
+@import models.Mode
+
+@(appConfig: FrontendAppConfig, form: Form[_], mode: Mode)(implicit request: Request[_], messages: Messages)
+
+@main_template(
+    title = s"${messages("messages__psa__name__title")}",
+    appConfig = appConfig,
+    bodyClasses = None) {
+
+    @helpers.form(action = PsaNameController.onSubmit(mode), 'autoComplete -> "off") {
+
+        @components.back_link()
+
+        @components.error_summary(form.errors)
+
+        @components.heading("messages__psa__name__heading")
+
+        <p>@messages("messages__psa__name__p1")</p>
+
+        @components.input_text(
+          field = form("psaName"),
+          label = messages("messages__psa__name__heading"),
+          labelClass=Some("visually-hidden")
+        )
+
+        @components.submit_button()
+    }
+}
